Add cast to avoid warning.
git-svn-id: http://svn.research.cc.gatech.edu/kaos/evpath/trunk@4665 8ba6a8b9-845b-0410-94f0-974be2f5df47

+++ cm.c
@@ -1784,7 +1784,7 @@ void *data;
 	 * their own buffers.  How to support?  Maybe do a copy and 
 	 * return the copy...?
 	 */
-	fprintf(stderr, "Error: INT_CMreturn_buffer called with record %lx not associated with cm\n", data);
+	fprintf(stderr, "Error: INT_CMreturn_buffer called with record %lx not associated with cm\n", (long)data);
     }
 }
